Add load-balancer heartbeat to float IP between load balancers
Primary LB runs as server, backup runs as client. If the primary
becomes unavailable for a short time (3s), backup claims the
IP and sends gratuitous arp to network. If the primary comes back
online, it tells the backup to relinquish the IP, then claims
the IP and sends gARP to inform network it's back.